+++ CHANGELOG.md
@@ -9,8 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Added cmdlet `Convert-ObjectToHashtable` wich can be used to convert any object
- to hashtable. It accepts objects from pipeline.
+- Added cmdlet `ConvertFrom-DscResourceInstance` wich can be used to convert any
+ object to in another format. It accepts objects from pipeline. [issue #71](https://github.com/dsccommunity/DscResource.Common/issues/71).
 
 ## [0.10.2] - 2021-03-24